removed typing errors
