+++ .gitignore
@@ -19,7 +19,7 @@ atlassian-ide-plugin.xml
 .settings
 .checkstyle
 .fbprefs
-bin
+
 # Automatically generated by WTP
 **/src/main/java/META-INF/
 **/src/test/java/META-INF/
