+++ xwiki-demo-cdk/bin/xwiki-demo-cdk.ts
@@ -18,25 +18,15 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-import * as cdk from '@aws-cdk/core'
-import { XwikiVpc } from '../lib/stacks/vpc'
-import { region } from '../lib/config'
-import { XwikiProductionStacks } from '../lib/stacks/xwiki-stacks'
+import 'source-map-support/register'
+import { App } from '@aws-cdk/core'
+import { EC2XWikiDemo } from '../lib/stacks/ec2-xwiki-demo'
+import { xwikidownload, region } from '../lib/stacks/config'
 
-const app = new cdk.App()
-
-const env = {
-  region: region
-}
-
-const xwikivpc = new XwikiVpc(app, 'xwiki-prod-vpc', {
-  env: env
-
+const app = new App()
+new EC2XWikiDemo(app, 'ec2XwikiDemo', {
+  xwiki: xwikidownload, //from the config file
+  env: {
+    region: region //from the config file
+  }
 })
-
-const xwikiproductionstacks = new XwikiProductionStacks(app, 'xwiki-ecs-loadbal', {
-  vpc: xwikivpc.xwikivpc,
-  env: env
-})
-
-xwikiproductionstacks.addDependency(xwikivpc)
